+++ eventstore/api/common/src/main/java/org/occurrent/eventstore/api/DuplicateCloudEventException.java
@@ -18,6 +18,7 @@ package org.occurrent.eventstore.api;
 
 import java.net.URI;
 import java.util.Objects;
+import java.util.StringJoiner;
 
 /**
  * An exception thrown if a cloud event already exists in the event store or if it violates some unique indexing rules.
@@ -25,11 +26,13 @@ import java.util.Objects;
 public class DuplicateCloudEventException extends RuntimeException {
     private final String id;
     private final URI source;
+    private final String details;
 
-    public DuplicateCloudEventException(String id, URI source, Throwable cause) {
-        super("Duplicate CloudEvent detected with id " + unknownIfNull(id) + " and source " + unknownIfNull(source == null ? null : source.toString()), cause);
+    public DuplicateCloudEventException(String id, URI source, String details, Throwable cause) {
+        super("Duplicate CloudEvent detected with id " + unknownIfNull(id) + " and source " + unknownIfNull(source == null ? null : source.toString()) + ", details: " + unknownIfNull(details), cause);
         this.id = id;
         this.source = source;
+        this.details = details;
     }
 
     public String getId() {
@@ -40,26 +43,30 @@ public class DuplicateCloudEventException extends RuntimeException {
         return source;
     }
 
+    public String getDetails() {
+        return details;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (!(o instanceof DuplicateCloudEventException)) return false;
         DuplicateCloudEventException that = (DuplicateCloudEventException) o;
-        return Objects.equals(id, that.id) &&
-                Objects.equals(source, that.source);
+        return Objects.equals(id, that.id) && Objects.equals(source, that.source) && Objects.equals(details, that.details);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, source);
+        return Objects.hash(id, source, details);
     }
 
     @Override
     public String toString() {
-        return "DuplicateCloudEventException{" +
-                "id='" + id + '\'' +
-                ", source=" + source +
-                '}';
+        return new StringJoiner(", ", DuplicateCloudEventException.class.getSimpleName() + "[", "]")
+                .add("id='" + id + "'")
+                .add("source=" + source)
+                .add("details='" + details + "'")
+                .toString();
     }
 
     private static String unknownIfNull(String str) {